+++ libraries/common/src/main/java/androidx/media3/common/audio/SonicAudioProcessor.java
@@ -33,7 +33,7 @@ import java.nio.ShortBuffer;
  * An {@link AudioProcessor} that uses the Sonic library to modify audio speed/pitch/sample rate.
  */
 @UnstableApi
-public class SonicAudioProcessor implements AudioProcessor {
+public final class SonicAudioProcessor implements AudioProcessor {
 
   /** Indicates that the output sample rate should be the same as the input. */
   public static final int SAMPLE_RATE_NO_CHANGE = -1;
@@ -100,7 +100,7 @@ public class SonicAudioProcessor implements AudioProcessor {
    *
    * @param speed The target factor by which playback should be sped up.
    */
-  public final void setSpeed(@FloatRange(from = 0f, fromInclusive = false) float speed) {
+  public void setSpeed(@FloatRange(from = 0f, fromInclusive = false) float speed) {
     checkArgument(speed > 0f);
     if (this.speed != speed) {
       this.speed = speed;
@@ -115,7 +115,7 @@ public class SonicAudioProcessor implements AudioProcessor {
    *
    * @param pitch The target pitch.
    */
-  public final void setPitch(@FloatRange(from = 0f, fromInclusive = false) float pitch) {
+  public void setPitch(@FloatRange(from = 0f, fromInclusive = false) float pitch) {
     checkArgument(pitch > 0f);
     if (this.pitch != pitch) {
       this.pitch = pitch;
@@ -131,7 +131,7 @@ public class SonicAudioProcessor implements AudioProcessor {
    * @param sampleRateHz The sample rate for output audio, in Hertz.
    * @see #configure(AudioFormat)
    */
-  public final void setOutputSampleRateHz(int sampleRateHz) {
+  public void setOutputSampleRateHz(int sampleRateHz) {
     checkArgument(sampleRateHz == SAMPLE_RATE_NO_CHANGE || sampleRateHz > 0);
     pendingOutputSampleRate = sampleRateHz;
   }
@@ -147,7 +147,7 @@ public class SonicAudioProcessor implements AudioProcessor {
    * @param playoutDuration The playout duration to scale.
    * @return The corresponding media duration, in the same units as {@code duration}.
    */
-  public final long getMediaDuration(long playoutDuration) {
+  public long getMediaDuration(long playoutDuration) {
     if (outputBytes >= MIN_BYTES_FOR_DURATION_SCALING_CALCULATION) {
       long processedInputBytes = inputBytes - checkNotNull(sonic).getPendingInputBytes();
       return outputAudioFormat.sampleRate == inputAudioFormat.sampleRate
@@ -172,7 +172,7 @@ public class SonicAudioProcessor implements AudioProcessor {
    * @param mediaDuration The media duration to scale.
    * @return The corresponding playout duration, in the same units as {@code mediaDuration}.
    */
-  public final long getPlayoutDuration(long mediaDuration) {
+  public long getPlayoutDuration(long mediaDuration) {
     if (outputBytes >= MIN_BYTES_FOR_DURATION_SCALING_CALCULATION) {
       long processedInputBytes = inputBytes - checkNotNull(sonic).getPendingInputBytes();
       return outputAudioFormat.sampleRate == inputAudioFormat.sampleRate
@@ -187,7 +187,7 @@ public class SonicAudioProcessor implements AudioProcessor {
   }
 
   /** Returns the number of bytes processed since last flush or reset. */
-  public final long getProcessedInputBytes() {
+  public long getProcessedInputBytes() {
     return inputBytes - checkNotNull(sonic).getPendingInputBytes();
   }
 
@@ -197,8 +197,7 @@ public class SonicAudioProcessor implements AudioProcessor {
   }
 
   @Override
-  public final AudioFormat configure(AudioFormat inputAudioFormat)
-      throws UnhandledAudioFormatException {
+  public AudioFormat configure(AudioFormat inputAudioFormat) throws UnhandledAudioFormatException {
     if (inputAudioFormat.encoding != C.ENCODING_PCM_16BIT) {
       throw new UnhandledAudioFormatException(inputAudioFormat);
     }
@@ -214,7 +213,7 @@ public class SonicAudioProcessor implements AudioProcessor {
   }
 
   @Override
-  public final boolean isActive() {
+  public boolean isActive() {
     return pendingOutputAudioFormat.sampleRate != Format.NO_VALUE
         && (shouldBeActiveWithDefaultParameters || !areParametersSetToDefaultValues());
   }
@@ -226,7 +225,7 @@ public class SonicAudioProcessor implements AudioProcessor {
   }
 
   @Override
-  public final void queueInput(ByteBuffer inputBuffer) {
+  public void queueInput(ByteBuffer inputBuffer) {
     if (!inputBuffer.hasRemaining()) {
       return;
     }
@@ -239,7 +238,7 @@ public class SonicAudioProcessor implements AudioProcessor {
   }
 
   @Override
-  public final void queueEndOfStream() {
+  public void queueEndOfStream() {
     // TODO(internal b/174554082): assert sonic is non-null here and in getOutput.
     if (sonic != null) {
       sonic.queueEndOfStream();
@@ -248,7 +247,7 @@ public class SonicAudioProcessor implements AudioProcessor {
   }
 
   @Override
-  public final ByteBuffer getOutput() {
+  public ByteBuffer getOutput() {
     @Nullable Sonic sonic = this.sonic;
     if (sonic != null) {
       int outputSize = sonic.getOutputSize();
@@ -272,12 +271,12 @@ public class SonicAudioProcessor implements AudioProcessor {
   }
 
   @Override
-  public final boolean isEnded() {
+  public boolean isEnded() {
     return inputEnded && (sonic == null || sonic.getOutputSize() == 0);
   }
 
   @Override
-  public final void flush() {
+  public void flush() {
     if (isActive()) {
       inputAudioFormat = pendingInputAudioFormat;
       outputAudioFormat = pendingOutputAudioFormat;
@@ -300,7 +299,7 @@ public class SonicAudioProcessor implements AudioProcessor {
   }
 
   @Override
-  public final void reset() {
+  public void reset() {
     speed = 1f;
     pitch = 1f;
     pendingInputAudioFormat = AudioFormat.NOT_SET;